Game vtbl extended, map init/free routines and reg_cfg game dtor update
game destructor now updates registry configuration values. This does not
have any effect since these changes are not written to disk. They are
not read from the disk on startup yet. The virtual table has been
extended with a couple of routines that are going to be called in
game_show_focus_screen.

map probably is not the best name for what it really does in the game.
It does some direct draw stuff and has some geometry objects for draw
boundaries. These geometry objects are not implemented yet.