tools/drs: shp type, drs mapping, list inspector, item mapping
drs now correctly shows the number of lists that are stored in a data
file. The drs_map structure needs to be extended because the number of
items in a list is not read correctly.

Some minimal error checking is provided for out of bounds access. The
checks are not fail-proof yet!