tools/xmap: data dir section mapping
Data directory is parsed and all different sections are printed with
their name. I've also ran a couple of tests and valgrind is happy and
bad files are parsed up till the point that it realises fields are very
wrong. We still need to consider if we want to break out early or
continue until we encounter a serious violation.