tools/xmap: sub tree stat loops func call fix, sub pos fix
Sub tree statistics loops for name and id entries erroneously called
rsrc_ststat when it should have called rsrc_tstat. The rdi_name_start
offset was computed properly, but the tree statistic funcs assumed that
the offset included the resource directory structure. For now, we just
don't add this to the offset. We need to think about a more elegant
solution.