tools/rsrc: parent resdir offset in rsrc_walk
We have to implement a relative offset that is passed through in order
to make sure we can traverse the node indices correctly.

String indices are computed as resdir_id=id/16+1 resdir_offset=id%16.
Keep in mind that the parent offset in our reference program, pcrtest,
is broken and the values should be ignored.