openal sfx hacked together
Sound effects audio playback now supports looping and a volume. The
sources and buffers for OpenAL are created once and reused for all other
effects.
The music lookup table has been downgraded to a simpler model since it
does not make sense to include all clips into it because we would get a
huge resource table lookup mess that we don't want.