Main executable redesign, code cleanup
The current design is workable but not very easy to maintain. Many
engine specific tasks are going to be moved to the genie directory which
should take of all core stuff anyway.

The default bitmap font is embedded by default because this is known to
cause problems if it is not run from the same directory which only makes
packaging the project worse.