dmap redesign, auto chdir, reload drs data
The legacy drs code is partially reused by redesigning the new drs
loading mechanism. It is easier to properly handle memory management.

Most new code has been written in C89 style, so we can backport it to
older machines if we eventually want to make it run on older devices as
well.