Reverse engineering tools
Tools that help reverse engineering or immitate the game's behavior are
going to be implemented in reverse_engineering/ as i686 PE32 executables
that are cross compiled on linux. One registry fetching tool has been
implemented.

These tools are optional to build and are therefore not included in the
main configure script.