reverse_engineering: run script, init improvements
Figured out a couple of variables and magic constants as well as
extended the setup initialization code and completed simple string error
handling.

Also provided a simple script that builds the aoe setup and runs it
directly in the CD-ROM setup image directory for rapid testing. A simple
usage is provided in the script (i.e. run_setup.sh) how to run it.

The language specific initilization code in init_lang uses some weird
logic with FindResourceA(3w) that we may want to further decipher into
their macro equivalents.

The main routine for setup.c has been completed, but we need to figure
out whether the annotated variables may be arguments to init.