reverse_engineering: hwnd script resource loading regression (wip)
It seems like the windows resources in the original aoesetup.exe are
required for proper startup. Unfortunately, we have to create a
workaround because I don't know how we can extract them easily and
update them in our port.

For now, the idea is to start working on rp (i.e. a resource patcher)
that copies the resources from the original setup into our setup.