reverse_engineering/setup: fix resource regression
Fixed regression in script resource regression language initialization
code preventing the AoE setup to enter the main event loop. A lot of
configure code had to be rewritten in order to reflect these changes:
the configure script has been removed entirely in reverse_engineering/
and a static Makefile has been provided to also deal with the .rc files
that are required for the language initialization code.

The resource patcher is somewhat working and appears to be stable, but I
think we will remove this is a later revision.