reverse_engineering/setup: main event handling (wip)
Add main loop and handle many events and log events that are not
completely ported yet. Also ignore .bin files that are generated by the
Makefile which should not be checked in for obvious reasons.

A simple sound cleanup hook has been added as well as some more global
variables and renamed a couple of them.

Finally, a simple panic handler has been implemented to improve error
handling and reducing duplicated code.