add cdrom support, ttf stub
The new CD-ROM subsystem checks if the game CD-ROM is inserted and
mounted properly. It also looks for a couple of files that are required
during startup. genie_game takes care of starting and stopping the
CD-ROM logic.

Initial True Type Font (=TTF) support is implemented but not used yet. A
couple of default font paths for the microsoft fonts provided through
wine are implemented. The gfx code takes care of starting and stopping
the TTF logic.

Finally, improved the image loading in reverse_engineering/setup.c.