general documentation updated, wine locating
The project's status amongst other things were seriously outdated in the
documentation. Some information from the issue tracker have been copied
to the documentation and HACKING has been completely revised.

The engine now determines if wine has been installed, because this may
indicate that the game is installed already in the user's wine
directory. It does not search for the installation path yet.

Lastly, added a simple check if the game is run as root and halts
execution to prevent wreaking havoc with root privileges.