basic console integration
A console can be accessed using ` and a couple of basic commands are
provided. Supported commands are: help info quit

The console shows up to 40 lines with a maximum of 256 characters per
line. Lines that are too long are not broken up yet. Newlines in text
is recognized and interpeted correctly. All text printed to the console
is also printed to stdout for testing and debugging purposes.

The input line is prepended with `# ' and can be changed quite easily in
CONSOLE_PS1 in genie/ui.c. Other limits for the console are defined in
genie/ui.h.