drs fetching from wine installation or CD-ROM
This patch relinquishes the end-user having to copy the executable to a
directory where the game data is located, because it is now able to find
them automatically. The cool thing is, the game does not have to be
installed through wine in order to make this work.

The CD-ROM initialization code has been moved from the genie game logic
to the genie engine logic, because the drs data must be fetched from
disk before the UI can be initialized. The CD-ROM code used to run after
the UI had been initialized which would make it impossible to work
without changing the order.

Finally, the shell did not properly handle invalid options provided to
the toggle command because it did not check for unknown options.