OpenAL audio support 2.0
The original OpenAL support is redesigned because the previous version
was a broken mess. This subsystem still has lots of old code inherited
from various sources in the project and we are going to refactor them
once we add more functionality that depends on the sfx subsystem.

Unfortunately, SDL2_mixer does not support different audio formats
played together and applying effects is not nearly as complete as with
OpenAL.

So, for the music tracks we will use SDL2_mixer because we don't need
any effects except fade-in/fade-out and OpenAL for the sound effects.

The wave parsing code is eventually going to be provided as a standalone
component for easy integration in other components or even other
projects.

The first noticeable result is in the main menu where button presses now
play the default button click sound. Some hacks are still residing in
the code (e.g. dmap) that needs to be fixed. The clip cleanup code is
not necessary and is therefore not provided yet.

Finally, changed the cdrom absolute path to include "game/" for normal
use.