Marked old executable as obsolete
The code is not going to be removed immediately because a lot of its
behavior will be replicated in the genie game engine code.

The root configure script now does not automatically compile the tools
because `normal' end-users will probably not care about them anyway.

From now on, the only headers that are allowed to be included outside
genie/ are located in genie/include/genie to make it easier to separate
the game engine.