video command line options, sfx heap fix, system wrapper
Many changes are crammed into this patch. The new command line options
are introduced:

  novideo   Disable all avi video playback
  nointro   Disable all logo and intro videos
  grpintro  Play multiple groups in one playlist.
            Skipping this will skip the entire playlist.

The game will hide when the intro and logo videos are played. When all
intro videos are played, the game will become visible again and request
focus.

The sfx heap did not correctly sift added elements which resulted in
fragmentation and this could even corrupt the heap.

Finally, introduced a wrapper for system(3) calls for debugging
purposes.