Crash handler with backtrace and debug info
If the game crashes, the default crash handler is invoked to show a
backtrace and some additional debug info that can be copy-pasted in a
bug report for easy debugging.

The configure scripts had to be changed in order to include symbol info
by default.

You can invoke the crash handler manually using the undocumented command
`hcf' (halt and catch fire). This only works when building in debug
mode!

You can disable the computation of the faulted address by defining
NO_FAULT_ADDRESS in crash.c