Single player menu's
Many new player menu's have been implemented for the real gameplay. The
menu logic is limited because it cannot render a menu onto another one.

The default in-game music is played if a new random map is started. The
defeat music will be played as well when the game is quit.