empires: add background music, auto detect music
If the audio tracks are also mounted on a iso9660 device, it is
recognised and played in game. The main menu, victory, defeat and first
in game tracks are played.

You can also mute music and/or sound effects with the command line
options provided in the official readme document.