basic viewport movement, dummy villager
the villager image is not drawn properly because the player color slp
commands are not recognised (this also applies to the buildings). most
desert tiles also have incomplete slp commands.

you can pause and unpause the game with F3. some glue logic has been
added that needs to be cleanup up and the render code has been
refactored a little bit.