empires: don't require wine installation, fix audio init
If the game wasn't installed with wine, it would not start because the
DRS files from the CD-ROM weren't used. Also, audio initialisation did
not work on Xubuntu 18, because SDL2 apparently doesn't support anything
other than wave...

Finally updated the drs code because it would not compile the I/O handling in
release mode.