World logic refactoring
All game units are now handled by the global game model and all units
are drawn in the world. The Quadtree logic is still very basic, but this
saves us time converting one system to the other.

Unit animation logic has been changed to reflect these changes. The
in-game viewport is roughly limited to the world's boundaries, but it is
not identical to the original game.