create ttf bitmap renderer
in order to finally fix the libfreetype issues, we have created a tool
that can visualize data created with reverse_engineering/font. this is
going to be a fairly painful process, but it is really worth it.

when the new ttf bitmap renderer is in place, the fonts are initially
not anti-aliased, because this is more difficult because we have to
recreate the alpha channel.