integrate xtcommon, cross compile 32 and 64 bits
played around with the font dumper tool and concluded that anti-aliasing
is currently too much work to integrate correctly. xtcommon is added
because we have some ugly hacks in the code to make it cross-platform,
and i don't want to deal with them anymore and just let xtcommon handle
that. it's one of the reasons it was created in the first place!