empiresx: restart with cmake
I have lost count how many refactorings I am doing or planning to do,
but this should really ease development (for the developers as well as
for end-users). The project builds automagically in VStudio 2019
Community edition with target x64-Debug. Linux should work, but it needs
a bit of OS dependent stuff to make it work properly.