add game tick stub, prepare match logic
if a match is started, the clients will notify the server when they are
ready. the server in its turn will announce all players and start the
game. i've tried to keep the magic handlers to a minimum, but the
dedicated and GUI server have minor changes for the prepare match logic,
but this should be the only real minor difference.

the user id has been changed to uint16_t (the same as player id) since
it does not make sense to have more than 65536 players. in theory, this
can mess things up when the server is up for a long time and accidently
reuses a player id that's already in use, but in practise, this should
never happen.