draw terrain trees, wrap image count, minor bugfixes
if an image was drawn with an index out of bounds, the game would crash.
specifying an image that is out of bounds sounds like a bug, but we
don't want this, since the dedicated server does not care about this and
the graphical client should just wrap the index if the animation keeps
looping. the player index definitely must be wrapped because we support
more than 9 players (including gaia).

some code was very intertwined with other logic that decreased overall
maintenance, so these have been partially or completely split into
separate files.

some other stuff:
* commented confirmation messages to avoid output clutter
* generalized math.hpp:tile_to_scr
* add start help to dedicated server
* extract vector to geom.hpp, viewport in empiresx.cpp