refactor image texture and terrain tile mapping
the graphics subsystem did not provide a simple mapping to get the
actual animation and subimages from the texture, which will create a lot
of problems for menus that need different drs assets.

scenario static objects like tiles are stored a to be determined layout
and order from the original game. to simplify the terrain logic and
visualisation, each tile now has a tile type identifier and a subimage,
which currently also uses a similar technique as described above.

ui has also been extended with a listpopup element (still wip). started
working on the singleplayer menu. text api calls now also compute the
final dimensions of the drawn text. this can be used for elements that
need to know the text size in advance.