fix host game timeout, make popups async
host client could not connect in time to server because the threadpool
was overpopulated. so when the server start thread died, the host client
thread would run. therefore, it could never connect. for now we use a
simple assumption that reserving before pushing the thread there will be
no time window where another thread overpopulates the threadpool.

also, i overlooked that pushing popups was racy. this has been addressed
by keeping an async copy that is copied when popups are processed in the
ui thread.

socket operations are also made more multi-thread friendly. i won't
claim it is race-proof, but it prevents using sockets that are modified
while a tcpsocket function is running. if it would change, it should
error out gracefully rather than try to send or receive data on the new
socket.

finally, enabled profiler runs for unit tests would break network tests
that assume no wsastartup has run yet. when tracy is running, these
tests are skipped and it will print that these tests are being skipped.