partial fix for network lagging when entering bogus ui input field values
so apparently, imgui does some internal bookmarking that we interpret as
spurious truth values when checking if it's scalar input field has been
modified. this messes up the change detection system sending spurious
scenario settings updates to the server. the initial spurious update
doesn't seem suppressed though, so it still fucks it up, but not as bad
as it only lags one network packet now.