disable eventloop thread, track peers, fix serversocketctl handling
since the eventloop is not used yet, there's no point to keep it
running. the server now registers incoming and dropped peers. also,
the serversocket would crash when shut down as the controller stopped
callback would be called twice. the second time, the controller was
already deconstructed. now it resets the controller callback after one
stopped call.

finally, lowered the maximum packet size to the real minimum, according
to the packet header layout on wikipedia.