broadcast incoming and dropped peers
clients will now know when peers join and leave. if a client joins while
other clients have already joined, the server won't notify the client
yet. this is something i would love to implement, but not close to
midnight as it's very likely to be buggy if i don't get some sleep now.