add ref to chat messages, share usernames to other peers
usernames are stored in client and passed to the engine when e.g. a
client sends a chat message. set_username and peermod::set_ref_username
are ambiguous now. we might drop set_username as it doesn't really add
value.

also disabled network logging for try_send and queue_out as it clutters
the output too much.