fix font loading, stub gameplay and upgrade hud
stub dummy map with a colorless town center where a camera can be moved
on the screen. the camera is controller frame independently and the game
keys are going to be configurable using a new introduced keyboard
controller.

font loading failed for debug builds as imgui asserts when it cannot be
loaded. the solution is an ugly hack as it provides a small time window
to trick the engine to load a different font if the file is changed just
at the right moment.

the in game chat can now be closed by clicking the chat button in the
main menu bar. the hud now uses the real game graphics for the menu bar
and the HUD in the bottom screen part.

also, slp parsing was still buggy. this should now be addressed. the
only byte code that isn't implemented is 0x0e. i'm not sure if AoE
uses this at all. AoE II does use it, but we don't need it afaik.

the eventloop thread has been moved from the engine to the server. it is
merely started, but no game logic is applied yet.

finally, maps are allowed to wrap, although no maps are generated yet.