introduce overlay tiles
overlay tiles are loaded, but only the grass desert one is being used.
terrain generation has been simplified to create a desert only map. as
the subimages are chosen randomly, there are no smooth transitions yet.
we need to look for adjacent tiles that are allowed in the original game.

also, the heightmap is still not being used.